% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_funcs.R
\name{scatter}
\alias{scatter}
\title{Consistent scatterplot graphing}
\usage{
scatter(df, x, y, group = NULL, alpha_v)
}
\arguments{
\item{df}{A dataframe}

\item{x}{A vector}

\item{y}{A vector}

\item{group}{A vector}
}
\value{
A scatterplot of \code{y} on \code{x}
}
\description{
Consistent scatterplot graphing
}
